dist-requirements.txt
build
twine
PyYAML==6.0
SQLAlchemy==1.4.35
tabulate==0.8.7
